+++ nutype_macros/src/common/models.rs
@@ -17,13 +17,17 @@ use super::gen::type_custom_closure;
 
 /// A spanned item. An item can be anything that cares a domain value.
 /// Keeping a span allows to throw good precise error messages at the validation stage.
-#[derive(Debug)]
+#[derive(Debug, Clone)]
 pub struct SpannedItem<T> {
     pub item: T,
     pub span: Span,
 }
 
 impl<T> SpannedItem<T> {
+    pub fn new(item: T, span: Span) -> Self {
+        Self { item, span }
+    }
+
     pub fn span(&self) -> Span {
         self.span
     }
@@ -405,3 +409,51 @@ impl ToTokens for TypedCustomFunction {
         };
     }
 }
+
+/// This trait allows to reuse validation of numeric validators.
+pub trait NumericBoundValidator<T: Clone> {
+    fn greater(&self) -> Option<T>;
+    fn greater_or_equal(&self) -> Option<T>;
+    fn less(&self) -> Option<T>;
+    fn less_or_equal(&self) -> Option<T>;
+}
+
+macro_rules! impl_numeric_bound_validator {
+    ($tp:ident) => {
+        impl<T: Clone> crate::common::models::NumericBoundValidator<T> for $tp<T> {
+            fn greater(&self) -> Option<T> {
+                if let $tp::Greater(value) = self {
+                    Some(value.clone())
+                } else {
+                    None
+                }
+            }
+
+            fn greater_or_equal(&self) -> Option<T> {
+                if let $tp::GreaterOrEqual(value) = self {
+                    Some(value.clone())
+                } else {
+                    None
+                }
+            }
+
+            fn less(&self) -> Option<T> {
+                if let $tp::Less(value) = self {
+                    Some(value.clone())
+                } else {
+                    None
+                }
+            }
+
+            fn less_or_equal(&self) -> Option<T> {
+                if let $tp::LessOrEqual(value) = self {
+                    Some(value.clone())
+                } else {
+                    None
+                }
+            }
+        }
+    };
+}
+
+pub(crate) use impl_numeric_bound_validator;